backend/pyproject.toml
[project]
name = "govflow-backend"
version = "0.1.0"
description = "GovFlow Agentic GraphRAG Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "gremlinpython>=3.7.0",
    "openai>=1.0",
    "oss2>=2.18",
    "boto3>=1.35",
    "asyncpg>=0.30.0",
    "websockets>=14.0",
    "python-jose[cryptography]>=3.3",
    "slowapi>=0.1.9",
    "pyyaml>=6.0",
    "httpx>=0.28.0",
    "jinja2>=3.1",
    "python-multipart>=0.0.18",
    "aiofiles>=24.0",
    "argon2-cffi>=23.1.0",
    "python-magic>=0.4.27",
    "werkzeug>=3.0",
    # 3.4 structured logging
    "python-json-logger>=2.0",
    # 3.6 OpenTelemetry tracing
    "opentelemetry-api>=1.28",
    "opentelemetry-sdk>=1.28",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-asyncpg>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28",
    # 3.7 Sentry
    "sentry-sdk[fastapi]>=2.18",
    # 3.8 Alembic
    "alembic>=1.14",
    # 3.5 Prometheus metrics
    "prometheus-fastapi-instrumentator>=7.0.0",
    # resilience: retry + circuit breaker helpers
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires Docker services (Gremlin, PostgreSQL, MinIO)",
    "requires_dashscope: requires DASHSCOPE_API_KEY env var for real LLM calls",
    "slow: takes >10 seconds to run",
    "benchmark: accuracy and latency benchmark tests",
]
